convert to classes. add modules.
