+++ submissions/Nemooochka/friends-app/index.html
@@ -55,5 +55,5 @@
     </div>
 </main>
 
-<script src="scripts.js"></script>
+<script src="js/app.js" type="module"></script>
 </body>
